fixed saving & loading of tool's fill property

+++ backend/Bru/src/main/java/BruGroup/Bru/Entity/Tool.java
@@ -19,16 +19,16 @@ public class Tool {
     private double y;
 
     private String name;
-    private int amount;
+    private double amount;
     private String color;
-    private int height;
-    private int width;
+    private double height;
+    private double width;
 
     public Tool() {
 
     }
 
-    public Tool(ToolIdentity toolIdentity, double x, double y, int amount, String color, String name, int width, int height) {
+    public Tool(ToolIdentity toolIdentity, double x, double y, double amount, String color, String name, double width, double height) {
         this.toolIdentity = toolIdentity;
         this.x = x;
         this.y = y;
@@ -71,11 +71,11 @@ public class Tool {
         this.y = y;
     }
 
-    public int getAmount() {
+    public double getAmount() {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
 
@@ -87,13 +87,13 @@ public class Tool {
         this.color = color;
     }
 
-    public int getHeight() {return height;}
+    public double getHeight() {return height;}
 
-    public int getWidth() {return width;}
+    public double getWidth() {return width;}
 
-    public void setHeight(int height) {this.height = height;}
+    public void setHeight(double height) {this.height = height;}
 
-    public void setWidth(int width) {this.width = width;}
+    public void setWidth(double width) {this.width = width;}
 
     public String getName() {
         return name;
